vul-scan2.2/requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
playwright>=1.30.0
pyjwt>=2.6.0
urllib3>=1.26.0


# Add to requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.0.0